requirements.txt
openai
pyyaml
pytest
argparse
